fix: use a single sign method
