--- src/contracts/evm/index.ts
@@ -1,10 +1,0 @@
-import * as ChainSignaturesContractABI from './ChainSignaturesContractABI'
-import * as errors from './errors'
-export * from './ChainSignaturesContract'
-
-const utils = {
-  ChainSignaturesContractABI,
-  errors,
-}
-
-export { utils }
