+++ src/contracts/index.ts
@@ -1,4 +1,3 @@
 export { ChainSignatureContract, type SignArgs } from './ChainSignatureContract'
 
 export * as near from './near'
-export * as evm from './evm'
