+++ src/contracts/near/index.ts
@@ -1,7 +1,6 @@
 import * as signAndSend from './signAndSend'
 import * as transaction from './transaction'
 export * from './ChainSignatureContract'
-export * from './ChainSignatureContractWallet'
 
 const utils = {
   transaction,
